requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.2
pyasn1==0.3.2
pyasn1-modules==0.0.11
python-dotenv==0.6.5
requests==2.18.4
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.13
urllib3==1.22
Werkzeug==0.12.2
